+++ CHANGELOG.md
@@ -2,3 +2,11 @@
 
 **Added**
 -added CRUD
+
+## v1.1.0 - 22 March 2022
+
+**Added**
+-added support extension video
+
+**Modified**
+-modified schema